Create separate integration tests for plugin and handler
